+++ webextension/data/js/init-templates.js
@@ -1,9 +1,8 @@
-import {init} from "./backgroundTheme.js";
-
 const templatesSource = window.templatesSource = new Map();
-templatesSource.set('backgroundTheme', '/data/js/backgroundTheme.mst');
-templatesSource.set('panelCheckedDataItem', '/data/js/panelCheckedDataItem.mst');
-templatesSource.set('panelRssLinks', '/data/js/panelRssLinks.mst');
+templatesSource.set('backgroundTheme', '/data/template/backgroundTheme.mst');
+templatesSource.set('panelCheckedDataItem', '/data/template/panel/checkedDataItem.mst');
+templatesSource.set('panelRssLinks', '/data/template/panel/rssLinks.mst');
+templatesSource.set('tabMover', '/data/template/panel/tabMover.mst');
 
 moment.locale(browser.i18n.getMessage('language'));
 
@@ -46,6 +45,5 @@ loadMustacheTemplates(templatesSource)
 	.then(async (loadMap)=>{
 		appGlobal.mustacheTemplates = loadMap;
 		await chromeSettings.loadingPromise;
-		init();
 	})
 ;