+++ webextension/data/js/init-zdk.js
@@ -1,13 +1,9 @@
 import {ZDK} from "./classes/ZDK.js";
-import {setBaseDir} from './classes/loadJS.js';
 
 // appGlobal: Accessible with browser.extension.getBackgroundPage();
 window.appGlobal = {};
 
-const ADDON_JS_ROOT = window.ADDON_JS_ROOT = '/data/js';
-setBaseDir(ADDON_JS_ROOT);
-
-const zDK = new ZDK(ADDON_JS_ROOT);
+const zDK = new ZDK();
 const backgroundPage = browser.extension.getBackgroundPage();
 if (backgroundPage !== null) {
 	backgroundPage.zDK = zDK;