+++ webextension/data/js/panel/panel-init.js
@@ -20,40 +20,73 @@ import('../browserDetect.js')
 	.catch(console.error)
 ;
 
-chrome.runtime.getBackgroundPage(_backgroundPage => {
-	backgroundPage = _backgroundPage;
+async function baseInit() {
+	if (typeof browser === 'undefined' || browser === null) {
+		await import('../lib/browser-polyfill.js');
+	}
+	await import('../options-api.js');
 
-	backgroundPage.baseRequiredPromise.then(() => {
-		applyPanelSize();
+	backgroundPage = await window.browser.runtime.getBackgroundPage();
+	window.appGlobal = backgroundPage.appGlobal;
+	await backgroundPage.baseRequiredPromise;
+}
+const baseInitPromise = baseInit();
+baseInitPromise.then(async () => {
+	applyPanelSize();
 
-		window.optionColorStylesheet = backgroundPage.backgroundTheme.theme_cache_update(document.querySelector('#generated-color-stylesheet'));
-		if (typeof optionColorStylesheet === 'object' && optionColorStylesheet !== null) {
-			console.info("Theme update");
+	const {theme_cache_update} = await import('../backgroundTheme.js');
+	window.optionColorStylesheet = await theme_cache_update(document.querySelector('#generated-color-stylesheet'));
+	if (typeof optionColorStylesheet === 'object' && optionColorStylesheet !== null) {
+		console.info("Theme update");
 
-			let currentThemeNode = document.querySelector('#generated-color-stylesheet');
-			currentThemeNode.parentNode.removeChild(currentThemeNode);
+		let currentThemeNode = document.querySelector('#generated-color-stylesheet');
+		currentThemeNode.parentNode.removeChild(currentThemeNode);
 
-			document.querySelector('body').dataset.theme = optionColorStylesheet.dataset.theme;
+		document.querySelector('body').dataset.theme = optionColorStylesheet.dataset.theme;
 
-			document.querySelector('head').appendChild(optionColorStylesheet);
-		}
+		document.querySelector('head').appendChild(optionColorStylesheet);
+	}
 
-		document.querySelector('#disableNotifications').classList.toggle('off', backgroundPage.appGlobal['notificationGlobalyDisabled']);
-		document.querySelector('#disableNotifications').dataset.translateTitle = (backgroundPage.appGlobal['notificationGlobalyDisabled'])? 'GloballyDisabledNotifications' : 'GloballyDisableNotifications';
-	});
+	document.querySelector('#disableNotifications').classList.toggle('off', backgroundPage.appGlobal['notificationGlobalyDisabled']);
+	document.querySelector('#disableNotifications').dataset.translateTitle = (backgroundPage.appGlobal['notificationGlobalyDisabled'])? 'GloballyDisabledNotifications' : 'GloballyDisableNotifications';
 });
 
 window.onload = function () {
 	window.onload = null;
-	let jsFiles = [];
-	if(typeof browser === 'undefined' || browser === null) {
-		jsFiles.push('/lib/browser-polyfill.js');
-	}
-	jsFiles = jsFiles.concat(['options-api.js', 'lib/lodash.custom.min.js', 'copyToClipboard.js', 'panel/panel.js']);
 
-	import('../classes/loadJS.js')
-		.then(({loadJS}) => {
-			loadJS(document, jsFiles);
-		})
-	;
+
+	const currentCat = window.localStorage.getItem('panel-current-cat');
+	if (currentCat !== null) {
+		/**
+		 *
+		 * @type {HTMLInputElement}
+		 */
+		const input = document.querySelector(`input#${currentCat}`);
+		input.checked = true;
+	}
+	document.addEventListener('change', function (e) {
+		const node = e.target.closest('input[type="radio"][name="sections"]');
+		if (!node) return;
+
+		window.localStorage.setItem('panel-current-cat', node.id);
+	});
+
+	(async () => {
+		await baseInitPromise;
+
+		let jsFiles = [
+			'../lib/lodash.custom.min.js',
+			'../copyToClipboard.js',
+			'../panel/browserTabUtils.js',
+			'../panel/tabMover.js',
+			'../panel/pwa.js',
+			'../panel/service-worker.js',
+			'../panel/panel.js'
+		];
+
+
+		for (let src of jsFiles) {
+			await import(src);
+		}
+	})();
 };