+++ webextension/data/js/variousFeatures/windowsContextMenu.js
@@ -3,22 +3,12 @@
 import {ZDK} from "../classes/ZDK.js";
 
 let browserWindows = appGlobal["windows"] = [],
-	windowSwitchContextMenu_contexts = [],
-	windowSwitchContextMenu_subMenu = new Map(),
 	linkWindowContextMenu_subMenu = new Map()
 ;
 
 
-if (browser.contextMenus.ContextType.hasOwnProperty("PAGE")) {
-	windowSwitchContextMenu_contexts.push(browser.contextMenus.ContextType.PAGE)
-}
-if (browser.contextMenus.ContextType.hasOwnProperty("TAB")) {
-	windowSwitchContextMenu_contexts.push(browser.contextMenus.ContextType.TAB)
-}
-
-
 async function getCurrentWindowIds() {
-	if (windowSwitchContextMenu === null || linkWindowContextMenu === null) {
+	if (linkWindowContextMenu === null) {
 		return;
 	}
 
@@ -33,20 +23,13 @@ async function getCurrentWindowIds() {
 		}
 	});
 
-	await browser.contextMenus.update(windowSwitchContextMenu, {
-		"enabled": browserWindows.length > 1,
-		"title": browserWindows.length > 1 ? i18ex._("move_tab_of_window") : i18ex._("no_other_window")
-	});
+	console.dir(browserWindows)
 
 	await browser.contextMenus.update(linkWindowContextMenu, {
 		"enabled": browserWindows.length > 1,
 		"title": browserWindows.length > 1 ? i18ex._("open_link_to_window") : i18ex._("no_other_window")
 	});
 
-	for (let [browserWindowId, subMenuId] of windowSwitchContextMenu_subMenu) {
-		await browser.contextMenus.remove(subMenuId);
-		windowSwitchContextMenu_subMenu.delete(browserWindowId);
-	}
 	for (let [browserWindowId, subMenuId] of linkWindowContextMenu_subMenu) {
 		await browser.contextMenus.remove(subMenuId);
 		linkWindowContextMenu_subMenu.delete(browserWindowId);
@@ -59,20 +42,6 @@ async function getCurrentWindowIds() {
 				"windowId": browserWindow.id,
 				"windowType": "normal"
 			});
-
-			windowSwitchContextMenu_subMenu.set(
-				browserWindow.id,
-				browser.contextMenus.create({
-					"id": "winTab_" + browserWindow.id,
-					"contexts": windowSwitchContextMenu_contexts,
-					"enabled": true,
-					"parentId": windowSwitchContextMenu,
-					"title": i18ex._("window_windowId", {
-						"windowId": browserWindow.id,
-						"window": ZDK.stringEllipse(activeTab.title, 25)
-					})
-				})
-			);
 
 			linkWindowContextMenu_subMenu.set(
 				browserWindow.id,
@@ -94,44 +63,6 @@ async function getCurrentWindowIds() {
 
 	await windowContextMenu_update();
 	return browserWindows;
-}
-
-
-async function windowSwitchContextMenu_onClick(info, tab) {
-	let action_windowId = null,
-		tabIsActive = tab.active
-	;
-
-
-	if (windowSwitchContextMenu_subMenu.size > 0) {
-		let browserWindowTarget = null;
-		windowSwitchContextMenu_subMenu.forEach((subMenuId, browserWindowId) => {
-			if (browserWindowTarget === null) {
-				if (subMenuId === info.menuItemId) {
-					browserWindowTarget = browserWindowId;
-				}
-			}
-		});
-
-		if (browserWindowTarget !== null) {
-			action_windowId = browserWindowTarget;
-		}
-	} else {
-		action_windowId = (browserWindows[0].id !== tab.windowId) ? browserWindows[0].id : browserWindows[1].id;
-	}
-
-	if (action_windowId === null) {
-		throw "Cound not get Window";
-	}
-
-	await browser.tabs.move(tab.id, {
-		"windowId": action_windowId,
-		"index": -1
-	});
-
-	await browser.tabs.update(tab.id, {
-		"active": tabIsActive
-	});
 }
 
 
@@ -180,24 +111,6 @@ async function windowContextMenu_update() {
 			windowTypes: ["normal"]
 		});
 
-		for (let [browserWindowId, subMenuId] of windowSwitchContextMenu_subMenu) {
-			const [activeTab] = await browser.tabs.query({
-				"active": true,
-				"windowId": browserWindowId,
-				"windowType": "normal"
-			});
-
-			if (activeTab !== undefined) {
-				await browser.contextMenus.update(subMenuId, {
-					"enabled": currentBrowserWindow.id !== browserWindowId,
-					"title": i18ex._("window_windowId", {
-						"windowId": browserWindowId,
-						"window": ZDK.stringEllipse(activeTab.title, 25)
-					})
-				});
-			}
-		}
-
 		for (let [browserWindowId, subMenuId] of linkWindowContextMenu_subMenu) {
 			const [activeTab] = await browser.tabs.query({
 				"active": true,
@@ -223,31 +136,8 @@ async function windowContextMenu_update() {
  *
  * @type {number|null}
  */
-let windowSwitchContextMenu = null,
-	linkWindowContextMenu = null
-;
+let linkWindowContextMenu = null;
 window.baseRequiredPromise.then(function () {
-	let contextMenuParams = {
-		"id": "winTab_no_others",
-		"contexts": windowSwitchContextMenu_contexts,
-		"enabled": false,
-		"icons": {
-			"16": "/data/images/ic_open_in_browser_black_24px.svg"
-		},
-		"title": i18ex._("no_other_window")
-	};
-	try {
-		windowSwitchContextMenu = browser.contextMenus.create(contextMenuParams);
-	} catch (err) {
-		if (err.toString().indexOf('icons') === -1) {
-			console.warn(err);
-		}
-
-		delete contextMenuParams.icons;
-		windowSwitchContextMenu = browser.contextMenus.create(contextMenuParams);
-	}
-
-
 	let linkContextMenuParams = {
 		"id": "winTab_link_no_others",
 		"contexts": [
@@ -270,17 +160,19 @@ window.baseRequiredPromise.then(function () {
 		linkWindowContextMenu = browser.contextMenus.create(linkContextMenuParams);
 	}
 
-	getCurrentWindowIds();
+	getCurrentWindowIds()
+		.catch(console.error)
+	;
 });
 chrome.contextMenus.onClicked.addListener(function (info, tab) {
-	if (info.menuItemId === 'winTab_no_others') {
-		return windowSwitchContextMenu_onClick(info, tab);
-	} else if (info.menuItemId === 'winTab_link_no_others') {
-		return linkWindowContextMenu_onClick(info, tab);
-	} else if (info.menuItemId.startsWith('winTab_')) {
-		return windowSwitchContextMenu_onClick(info, tab);
+	if (info.menuItemId === 'winTab_link_no_others') {
+		return linkWindowContextMenu_onClick(info, tab)
+			.catch(console.error)
+		;
 	} else if (info.menuItemId.startsWith('winTab_link_')) {
-		linkWindowContextMenu_onClick(info, tab);
+		linkWindowContextMenu_onClick(info, tab)
+			.catch(console.error)
+		;
 	}
 });
 
@@ -291,6 +183,8 @@ browser.windows.onFocusChanged.addListener(windowContextMenu_update);
 browser.tabs.onUpdated.addListener(function (info, changeInfo, tab) {
 	if (tab.active === true && ((changeInfo.hasOwnProperty("status") && changeInfo.status === "complete") || changeInfo.hasOwnProperty("title"))) {
 		// Only update context menu if the active tab have a "complete" load
-		windowContextMenu_update.call(this, tab.windowId);
+		windowContextMenu_update.call(this, tab.windowId)
+			.catch(console.error)
+		;
 	}
 });